refactor: add docstrings and cleanup controller layer

+++ src/dioptra/restapi/group_membership/service.py
@@ -17,7 +17,7 @@
 """The server-side functions that perform group membership endpoint operations."""
 from __future__ import annotations
 
-from typing import List
+from typing import Any, cast
 
 import structlog
 from sqlalchemy.exc import IntegrityError
@@ -25,48 +25,56 @@ from structlog.stdlib import BoundLogger
 
 from dioptra.restapi.app import db
 
+from .errors import GroupMembershipDoesNotExistError
 from .model import GroupMembership
 
 LOGGER: BoundLogger = structlog.stdlib.get_logger()
 
 
 class GroupMembershipService(object):
-    @staticmethod
-    def create(
-        group_id: int,
-        user_id: int,
-        read: bool,
-        write: bool,
-        share_read: bool,
-        share_write: bool,
-        **kwargs,
-    ) -> GroupMembership:
-        log: BoundLogger = kwargs.get("log", LOGGER.new())  # noqa: F841\
+    def get_all(self, **kwargs) -> list[GroupMembership]:
+        """Retrieve a list of all group memberships.
 
-        return GroupMembership(
-            group_id=group_id,
-            user_id=user_id,
-            read=read,
-            write=write,
-            share_read=share_read,
-            share_write=share_write,
-        )
-
-    @staticmethod
-    def get_all(**kwargs) -> List[GroupMembership]:
+        Returns:
+            List[GroupMembership]: List of group memberships.
+        """
         log: BoundLogger = kwargs.get("log", LOGGER.new())  # noqa: F841
 
         return GroupMembership.query.all()  # type: ignore
 
-    @staticmethod
-    def get_by_id(group_id: int, user_id: int, **kwargs) -> GroupMembership | None:
+    def get(
+        self, group_id: int, user_id: int, error_if_not_found: bool = False, **kwargs
+    ) -> GroupMembership | None:
+        """Retrieve a group membership.
+
+        Args:
+            group_id (int): The unique ID of the group.
+            user_id (int): The unique ID of the user.
+            error_if_not_found (bool): Flag to raise an error if the membership is
+                not found.
+
+        Returns:
+            GroupMembership | None: The group membership if found, else None.
+
+        Raises:
+            GroupMembershipNotFoundError: If the membership is not found and
+                error_if_not_found is True.
+        """
+
         log: BoundLogger = kwargs.get("log", LOGGER.new())  # noqa: F841
 
-        return GroupMembership.query.filter(  # type: ignore
+        membership = GroupMembership.query.filter(
             GroupMembership.user_id == user_id, GroupMembership.group_id == group_id
         ).first()
 
-    def submit(
+        if error_if_not_found:
+            if membership is None:
+                log.error("Group Membership not found", group_id=group_id)
+                raise GroupMembershipDoesNotExistError
+
+        return cast(GroupMembership, membership)
+
+    def create(
         self,
         group_id: int,
         user_id: int,
@@ -76,10 +84,35 @@ class GroupMembershipService(object):
         share_write: bool,
         **kwargs,
     ) -> GroupMembership:
+        """Create a new group membership.
+
+        Args:
+            group_id (int): The unique ID of the group.
+            user_id (int): The unique ID of the user.
+            read (bool): Permission flag for read access.
+            write (bool): Permission flag for write access.
+            share_read (bool): Permission flag for sharing with read access.
+            share_write (bool): Permission flag for sharing with write access.
+
+        Returns:
+            GroupMembership: The newly created group membership.
+
+        Raises:
+            GroupMembershipAlreadyExistsError: If the membership already exists.
+        """
         log: BoundLogger = kwargs.get("log", LOGGER.new())
 
-        new_group_membership: GroupMembership = self.create(
-            group_id, user_id, read, write, share_read, share_write, log=log
+        if self.get(group_id=group_id, user_id=user_id) is not None:
+            log.error("Group Membership already exists", group_id=group_id)
+            raise GroupMembershipDoesNotExistError
+
+        new_group_membership = GroupMembership(
+            group_id=group_id,
+            user_id=user_id,
+            read=read,
+            write=write,
+            share_read=share_read,
+            share_write=share_write,
         )
 
         db.session.add(new_group_membership)
@@ -93,14 +126,30 @@ class GroupMembershipService(object):
 
         return new_group_membership
 
-    def delete(self, group_id, user_id, **kwargs) -> bool:
-        membership = self.get_by_id(group_id=group_id, user_id=user_id)
+    def delete(self, group_id, user_id, **kwargs) -> dict[str, Any]:
+        """Delete a group membership.
 
+        Args:
+            group_id: The unique ID of the group.
+            user_id: The unique ID of the user.
+
+        Returns:
+            A dictionary with the status and IDs of the deleted membership.
+        Raises:
+            IntegrityError: If there is an issue with the database integrity during
+                deletion.
+        """
+
+        log: BoundLogger = kwargs.get("log", LOGGER.new())
+
+        if (membership := self.get(group_id=group_id, user_id=user_id)) is None:
+            return {"status": "Success", "id": []}
         try:
             db.session.delete(membership)
             db.session.commit()
 
-            return True
+            log.info("Group Membership deleted", group_id=group_id, user_id=user_id)
+            return {"status": "Success", "id": [group_id, user_id]}
         except IntegrityError:
             db.session.rollback()
-            return False
+            return {"status": "Failure", "id": [group_id, user_id]}
